Lift fetch using SWR from HomePage to App
